SWER-117: optimize product visibility

+++ src/Persistor/ProductVisibilityPersistor.php
@@ -29,10 +29,14 @@ class ProductVisibilityPersistor
         $dto = new ProductVisibilityDTO($sku, $salesChannelIds);
         $dto = $this->transformer->transform($dto, $context);
 
-        $created = $this->productVisibilityRepository->create($dto->getCreatePayload(), $context);
+        if(!empty($dto->getCreatePayload())) {
+            $created = $this->productVisibilityRepository->create($dto->getCreatePayload(), $context);
+        }
 
-        $this->productVisibilityRepository->delete($dto->getDeletePayload(), $context);
+        if (!empty($dto->getDeletePayload())) {
+            $this->productVisibilityRepository->delete($dto->getDeletePayload(), $context);
+        }
 
-        return $created->getPrimaryKeys(ProductVisibilityDefinition::ENTITY_NAME);
+        return isset($created) ? $created->getPrimaryKeys(ProductVisibilityDefinition::ENTITY_NAME) : [];
     }
 }
